requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
frozenlist==1.3.1
geoip2==4.6.0
idna==3.3
lxml==4.9.1
maxminddb==2.2.0
multidict==6.0.2
pycparser==2.21
pyOpenSSL==22.0.0
requests==2.28.1
scapy==2.4.5
simplejson==3.17.6
soupsieve==2.3.2.post1
typing_extensions==4.3.0
urllib3==1.26.12
yarl==1.8.1
